extract reporting code to separate module
